requirements.txt
alembic==1.6.3
aniso8601==9.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.4
flask-marshmallow==0.14.0
Flask-Migrate==3.0.0
flask-restx==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
urllib3==1.26.4
Werkzeug==1.0.1
